Wrap radio buttons in a container
This wraps the items within the radio buttons component inside a
container and then optionally provides the ability to specify a
margin_bottom for this container.

It utilises prior art that was defined on button component within
static:
https://govuk-static.herokuapp.com/component-guide/button/with_margin_bottom

However the same mixin class is not available in
govuk_publishing_components so the code is reproduced here.